Move `set_datetime_now` fixture to a `conftest.py`
