+++ datalad_registry/tests/test_tasks/conftest.py
@@ -1,3 +1,5 @@
+from datetime import datetime, timezone
+
 import pytest
 
 from datalad_registry.models import RepoUrl, db
@@ -54,3 +56,19 @@ def processed_ds_urls(flask_app, two_files_ds_annex) -> list[int]:
         process_dataset_url(url_id)
 
     return url_ids
+
+
+@pytest.fixture
+def set_datetime_now(monkeypatch):
+    """
+    Set the `datalad_registry.tasks.datetime.now()` to a fixed value
+    """
+
+    class MockDateTime:
+        @classmethod
+        def now(cls, *_args, **_kwargs):
+            return datetime(2023, 9, 30, 19, 20, 34, tzinfo=timezone.utc)
+
+    from datalad_registry import tasks
+
+    monkeypatch.setattr(tasks, "datetime", MockDateTime)
